requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.6.10
django-kronos==0.6
django-toolbelt==0.0.1
django-widget-tweaks==1.3
djangorestframework==3.0.5
gunicorn==19.2.1
nose==1.3.4
Pillow==2.7.0
psycopg2==2.6
static3==0.5.1
